Update Info 	- version 3.275 Beta 	- Add Infra 		- DMSImpl 			- DMS Group in same location 	- Fixed Infra 		- Fixed Infra Class 			- Added setDMS 				- organize dms infrastructure 		- Fixed Infra Config 			- Added 'Load DMS XML Data' 		- Fixed InfraProperty 			- for DMS XML DATA 				- description 				- width_pixels 				- height_pixels 		- Fixed section 			- Fixed OrganizedStation 				- add DMS organize 			- Added Organized DMS 		- Fixed InfraType 			- Added DMS 	- Fixed InfraObject 		- Fixed Station 			- Added Upstream and DownStream DMS 			- Added Upstream and DownStream DMS Distance 			- Added MilePoint Distance 	- Added VSL Simulation 		- it is not available now 		- testing.. 	- Added TICAS Chart 		- testing 	- Fixed VISSIM Controll Error 		- Fixed Detector and DataCollection Error

+++ src/edu/umn/natsrl/infra/InfraObject.java
@@ -183,7 +183,7 @@ public abstract class InfraObject implements Serializable {
         return this.infraType;
     }
 
-    protected int calculateDistance(int e1, int e2, int n1, int n2)
+    protected int calculateDistance(double e1, double e2, double n1, double n2)
     {
         return (int) (Math.sqrt((e1 - e2) * (e1 - e2) + (n1 - n2) * (n1 - n2)) / 1609 * 5280);
     }
